+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 ## Unreleased
 ### Added
 - CORS configuration can be overwritten via settings
+- Pagination default page size can be set via settings
 
 ## 1.0.0-rc - 2018-03-22
 Initial release.